requirements_0524.txt
accelerate==0.29.2
aiohttp==3.9.3
aiohttp-retry==2.8.3
aiosignal==1.3.1
amqp==5.2.0
annotated-types==0.6.0
antlr4-python3-runtime==4.9.3
anyio==4.3.0
appdirs==1.4.4
asttokens==2.4.1
async-timeout==4.0.3
asyncssh==2.14.2
atpublic==4.1.0
attrs==23.2.0
auto_gptq==0.7.1
backcall==0.2.0
backports.zoneinfo==0.2.1
billiard==4.2.0
cachetools==5.3.3
celery==5.3.6
certifi==2024.2.2
cffi==1.16.0
charset-normalizer==3.3.2
click==8.1.7
click-didyoumean==0.3.1
click-plugins==1.1.1
click-repl==0.3.0
colorama==0.4.6
coloredlogs==15.0.1
comm==0.2.2
configobj==5.0.8
cryptography==42.0.5
dataclasses-json==0.6.4
datasets==2.18.0
debugpy==1.8.1
decorator==5.1.1
dictdiffer==0.9.0
dill==0.3.8
diskcache==5.6.3
distro==1.9.0
dpath==2.1.6
dulwich==0.21.7
dvc==3.42.0
dvc-data==3.8.0
dvc-gdrive==3.0.1
dvc-http==2.32.0
dvc-objects==3.0.6
dvc-render==1.0.1
dvc-studio-client==0.20.0
dvc-task==0.4.0
entrypoints==0.4
exceptiongroup==1.2.1
executing==2.0.1
fastapi==0.110.2
filelock==3.13.3
flatten-dict==0.4.2
flufl.lock==7.1.1
frozenlist==1.4.1
fsspec
funcy==2.0
gekko==1.1.1
gitdb==4.0.11
GitPython==3.1.43
google-api-core==2.18.0
google-api-python-client==2.126.0
google-auth==2.29.0
google-auth-httplib2==0.2.0
googleapis-common-protos==1.63.0
grandalf==0.8
greenlet==3.0.3
gto==1.6.2
h11==0.14.0
httpcore==1.0.5
httplib2==0.22.0
httpx==0.27.0
huggingface-hub==0.22.2
humanfriendly==10.0
hydra-core==1.3.2
idna==3.6
importlib_metadata==7.1.0
importlib_resources==6.4.0
ipykernel==6.29.4
ipython==8.12.3
iterative-telemetry==0.0.8
jedi==0.19.1
Jinja2==3.1.3
jsonpatch==1.33
jsonpointer==2.4
jupyter_client==8.6.1
jupyter_core==5.7.2
kombu==5.3.6
langchain==0.1.16
langchain-cli==0.0.21
langchain-community==0.0.34
langchain-core==0.1.45
langchain-text-splitters==0.0.1
langserve==0.1.0
langsmith==0.1.49
load-dotenv==0.1.0
markdown-it-py==3.0.0
MarkupSafe==2.1.5
marshmallow==3.21.1
matplotlib-inline==0.1.6
mdurl==0.1.2
mpmath==1.3.0
multidict==6.0.5
multiprocess==0.70.16
mypy-extensions==1.0.0
nest-asyncio==1.6.0
networkx==3.1
numpy==1.24.4
nvidia-cublas-cu12==12.1.3.1
nvidia-cuda-cupti-cu12==12.1.105
nvidia-cuda-nvrtc-cu12==12.1.105
nvidia-cuda-runtime-cu12==12.1.105
nvidia-cudnn-cu12==8.9.2.26
nvidia-cufft-cu12==11.0.2.54
nvidia-curand-cu12==10.3.2.106
nvidia-cusolver-cu12==11.4.5.107
nvidia-cusparse-cu12==12.1.0.106
nvidia-nccl-cu12==2.19.3
nvidia-nvjitlink-cu12==12.4.127
nvidia-nvtx-cu12==12.1.105
oauth2client==4.1.3
omegaconf==2.3.0
optimum==1.18.1
orjson==3.10.0
packaging==23.2
pandas==2.0.3
parso==0.8.3
pathspec==0.12.1
peft==0.10.0
pexpect==4.9.0
pickleshare==0.7.5
platformdirs==3.11.0
prompt-toolkit==3.0.43
proto-plus==1.23.0
protobuf==4.25.3
psutil==5.9.8
ptyprocess==0.7.0
pure-eval==0.2.2
pyarrow==15.0.2
pyarrow-hotfix==0.6
pyasn1==0.6.0
pyasn1_modules==0.4.0
pycparser==2.22
pydantic==2.6.4
pydantic_core==2.16.3
pydot==2.0.0
PyDrive2==1.19.0
pygit2==1.13.3
Pygments==2.17.2
pygtrie==2.5.0
pyOpenSSL==24.1.0
pyparsing==3.1.2
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
pytz==2024.1
PyYAML==6.0.1
pyzmq==25.1.2
regex==2023.12.25
requests==2.31.0
rich==13.7.1
rouge==1.0.1
rsa==4.9
ruamel.yaml==0.18.6
ruamel.yaml.clib==0.2.8
safetensors==0.4.2
scmrepo==2.0.4
semver==3.0.2
sentencepiece==0.2.0
shellingham==1.5.4
shortuuid==1.0.13
shtab==1.7.1
six==1.16.0
smmap==5.0.1
sniffio==1.3.1
SQLAlchemy==2.0.29
sqltrie==0.11.0
sse-starlette==1.8.2
stack-data==0.6.3
starlette==0.37.2
sympy==1.12
tabulate==0.9.0
tenacity==8.2.3
tokenizers==0.15.2
tomlkit==0.12.4
torch==2.2.2
tornado==6.4
tqdm==4.66.2
traitlets==5.14.2
transformers==4.39.3
triton==2.2.0
typer==0.9.4
typing-inspect==0.9.0
typing_extensions==4.10.0
tzdata==2024.1
uritemplate==4.1.1
urllib3==2.2.1
uvicorn==0.23.2
vine==5.1.0
voluptuous==0.14.2
wcwidth==0.2.13
xxhash==3.4.1
yarl==1.9.4
zc.lockfile==3.0.post1
zipp==3.18.1
click~=7.0
Sphinx
coverage
awscli
flake8
python-dotenv>=0.5.1
setuptools~=68.2.0
uvicorn~=0.29.0
fastapi~=0.110.0
elasticsearch~=8.12.1
aiostream~=0.5.2
langchain-elasticsearch~=0.1.1
langchain~=0.1.13
classy-fastapi~=0.6.1
huggingface-hub~=0.22.2
transformers~=4.39.0
aiohttp==3.9.3
aiohttp-retry==2.8.3
aiosignal==1.3.1
amqp==5.2.0
annotated-types==0.6.0
antlr4-python3-runtime==4.9.3
appdirs==1.4.4
asttokens==2.4.1
async-timeout==4.0.3
asyncssh==2.14.2
atpublic==4.1.0
attrs==23.2.0
backcall==0.2.0
backports.zoneinfo==0.2.1
billiard==4.2.0
celery==5.3.6
certifi==2024.2.2
cffi==1.16.0
charset-normalizer==3.3.2
click==8.1.7
click-didyoumean==0.3.1
click-plugins==1.1.1
click-repl==0.3.0
colorama==0.4.6
comm==0.2.2
configobj==5.0.8
cryptography==42.0.5
debugpy==1.8.1
decorator==5.1.1
dictdiffer==0.9.0
diskcache==5.6.3
distro==1.9.0
dpath==2.1.6
dulwich==0.21.7
dvc==3.42.0
dvc-data==3.8.0
dvc-http==2.32.0
dvc-objects==3.0.6
dvc-render==1.0.1
dvc-studio-client==0.20.0
dvc-task==0.4.0
entrypoints==0.4
executing==2.0.1
filelock==3.13.3
flatten-dict==0.4.2
flufl.lock==7.1.1
frozenlist==1.4.1
fsspec==2024.3.1
funcy==2.0
gitdb==4.0.11
GitPython==3.1.43
grandalf==0.8
gto==1.6.2
hydra-core==1.3.2
idna==3.6
importlib_metadata==7.1.0
importlib_resources==6.4.0
ipykernel==6.29.4
ipython==8.12.3
iterative-telemetry==0.0.8
jedi==0.19.1
jupyter_client==8.6.1
jupyter_core==5.7.2
kombu==5.3.6
markdown-it-py==3.0.0
matplotlib-inline==0.1.6
mdurl==0.1.2
multidict==6.0.5
nest-asyncio==1.6.0
networkx==3.1
omegaconf==2.3.0
orjson==3.10.0
packaging==24.0
parso==0.8.3
pathspec==0.12.1
pexpect==4.9.0
pickleshare==0.7.5
platformdirs==3.11.0
prompt-toolkit==3.0.43
psutil==5.9.8
ptyprocess==0.7.0
pure-eval==0.2.2
pycparser==2.22
pydantic==2.6.4
pydantic_core==2.16.3
pydot==2.0.0
pygit2==1.13.3
Pygments==2.17.2
pygtrie==2.5.0
pyparsing==3.1.2
python-dateutil==2.9.0.post0
PyYAML==6.0.1
pyzmq==25.1.2
requests==2.31.0
rich==13.7.1
ruamel.yaml==0.18.6
ruamel.yaml.clib==0.2.8
scmrepo==2.0.4
semver==3.0.2
shellingham==1.5.4
shortuuid==1.0.13
shtab==1.7.1
six==1.16.0
smmap==5.0.1
sqltrie==0.11.0
stack-data==0.6.3
tabulate==0.9.0
tomlkit==0.12.4
tornado==6.4
tqdm==4.66.2
traitlets==5.14.2
typer==0.12.1
typing_extensions==4.10.0
tzdata==2024.1
urllib3==2.2.1
vine==5.1.0
voluptuous==0.14.2
wcwidth==0.2.13
yarl==1.9.4
zc.lockfile==3.0.post1
zipp==3.18.1